Student Sleep_Patterns_&_Face_Happiness/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorflow
keras
flask
h5py
pillow
opencv-python 